yes stuff

+++ Projectiles/NPCProj/TheGuardian/GuardianPortalSwirl.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
@@ -9,63 +10,61 @@ namespace ElementsAwoken.Projectiles.NPCProj.TheGuardian
 {
     public class GuardianPortalSwirl : ModProjectile
     {
-        public int shrink = 100;
-
         public bool shrinking = true;
+        public float aiTimer = 100;
         public override void SetDefaults()
         {
-            projectile.width = 4;
-            projectile.height = 4;
+            projectile.width = 14;
+            projectile.height = 14;
+
             projectile.hostile = true;
-            projectile.ignoreWater = true;
             projectile.tileCollide = false;
-            projectile.alpha = 255;
+
             projectile.penetrate = -1;
-            projectile.timeLeft = 4000;
-            projectile.extraUpdates = 2;
+            projectile.timeLeft = 1000;
+
+            ProjectileID.Sets.TrailCacheLength[projectile.type] = 8;
+            ProjectileID.Sets.TrailingMode[projectile.type] = 0;
         }
         public override void SetStaticDefaults()
         {
             DisplayName.SetDefault("The Guardian");
         }
+        public override bool CanHitPlayer(Player target)
+        {
+            if (projectile.alpha >= 60) return false;
+            return base.CanHitPlayer(target);
+        }
         public override void AI()
         {
-            if (shrinking)
-            {
-                shrink -= 1;
-            }
-            if (!shrinking)
-            {
-                shrink += 1;
-            }
+            Lighting.AddLight(projectile.Center, 0.9f, 0.2f, 0.4f);
             Projectile parent = Main.projectile[(int)projectile.ai[1]];
-            int distance = shrink;
+            if (!parent.active) projectile.Kill();
+            aiTimer++;
+            float distance = 50 * (float)(1 + Math.Sin(aiTimer / 10f));
             double rad = projectile.ai[0] * (Math.PI / 180);
             projectile.position.X = parent.Center.X - (int)(Math.Cos(rad) * distance) - projectile.width / 2;
             projectile.position.Y = parent.Center.Y - (int)(Math.Sin(rad) * distance) - projectile.height / 2;
-            if (!parent.active)
-            {
-                projectile.Kill();
-            }
 
-            if (shrink <= 0)
+            projectile.alpha = parent.alpha;
+            if (Main.rand.NextBool(6))
             {
-                shrinking = false;
+                int dust = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, 6, projectile.velocity.X * 1.2f, projectile.velocity.Y * 1.2f, 130, default(Color), 2.75f);
+                Main.dust[dust].velocity *= 0.1f;
+                Main.dust[dust].scale *= 0.6f;
+                Main.dust[dust].noGravity = true;
             }
-            if (shrink >= 100)
-            {
-                shrinking = true;
-            }
-
-            int dust = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, 6, projectile.velocity.X * 1.2f, projectile.velocity.Y * 1.2f, 130, default(Color), 2.75f);
-            Main.dust[dust].velocity *= 0.1f;
-            Main.dust[dust].scale *= 0.6f;
-            Main.dust[dust].noGravity = true;
-
         }
-        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+        public override bool PreDraw(SpriteBatch sb, Color lightColor)
         {
-            target.immune[projectile.owner] = 2;
+            Vector2 drawOrigin = new Vector2(Main.projectileTexture[projectile.type].Width * 0.5f, projectile.height * 0.5f);
+            for (int k = 0; k < projectile.oldPos.Length; k++)
+            {
+                Vector2 drawPos = projectile.oldPos[k] - Main.screenPosition + drawOrigin + new Vector2(0f, projectile.gfxOffY);
+                Color color = projectile.GetAlpha(lightColor) * ((float)(projectile.oldPos.Length - k) / (float)projectile.oldPos.Length);
+                sb.Draw(Main.projectileTexture[projectile.type], drawPos, null, color, projectile.rotation, drawOrigin, projectile.scale, SpriteEffects.None, 0f);
+            }
+            return true;
         }
     }
 }
